Organized files into packages for better dev experience
