+++ nbproject/project.properties
@@ -71,7 +71,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=LudusUltimateTicTacToe
-main.class=ludusultimatetictactoe.UltimateTicTacToeFrame
+main.class=ultimate_tic_tac_toe.UltimateTicTacToeFrame
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
