--- src/ludusultimatetictactoe/Board.java
@@ -1,12 +1,0 @@
-// TODO
-
-
-package ludusultimatetictactoe;
-
-/**
- *
- * @author Ludus Team
- */
-public class Board {
-    
-}
