--- src/ludusultimatetictactoe/LudusUltimateTicTacToe.java
@@ -1,25 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Main.java to edit this template
- */
-package ludusultimatetictactoe;
-
-/**
- *
- * @author gaudium
- */
-public class LudusUltimateTicTacToe {
-
-    /**
-     * @param args the command line arguments
-     */
-    /*
-    public static void main(String[] args) {
-        System.out.println("Welcome to Ludus!");
-        System.out.println("Something something something");
-        System.out.println("testing git stuff");
-        System.out.println("testing again");
-    }
-    */
-    
-}
