+++ src/ultimate_tic_tac_toe/UltTTT.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package ludusultimatetictactoe;
+package ultimate_tic_tac_toe;
 import javax.swing.JButton;
 
 /**
