+++ src/ultimate_tic_tac_toe/UltimateTicTacToeFrame.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
  */
-package ludusultimatetictactoe;
+package ultimate_tic_tac_toe;
 
 import java.awt.Color;
 import java.awt.event.ActionEvent;
